Rename ressources to resources
