+++ run.py
@@ -1,3 +1,3 @@
-from ressources import client, get_config
+from resources import client, get_config
 
 client.run(get_config().BOT_TOKEN)
